Reorganize EV admin models and cleanup menu
