+++ ocpp/admin.py
@@ -16,9 +16,6 @@ from .models import (
     MeterValue,
     Transaction,
     Location,
-    ElectricVehicle,
-    Brand,
-    EVModel,
 )
 from .simulator import ChargePointSimulator
 from . import store
@@ -26,10 +23,7 @@ from .transactions_io import (
     export_transactions,
     import_transactions as import_transactions_data,
 )
-from core.admin import RFIDAdmin
-from core.models import WMICode
 from core.user_data import EntityModelAdmin
-from .models import RFID
 
 
 class LocationAdminForm(forms.ModelForm):
@@ -397,42 +391,3 @@ class MeterValueAdmin(EntityModelAdmin):
     )
     date_hierarchy = "timestamp"
     list_filter = ("charger", MeterValueDateFilter)
-
-
-@admin.register(ElectricVehicle)
-class ElectricVehicleAdmin(EntityModelAdmin):
-    list_display = ("vin", "license_plate", "brand", "model", "account")
-    search_fields = (
-        "vin",
-        "license_plate",
-        "brand__name",
-        "model__name",
-        "account__name",
-    )
-    fields = ("account", "vin", "license_plate", "brand", "model")
-
-
-class WMICodeInline(admin.TabularInline):
-    model = WMICode
-    extra = 0
-
-
-@admin.register(Brand)
-class BrandAdmin(EntityModelAdmin):
-    fields = ("name",)
-    list_display = ("name", "wmi_codes_display")
-    inlines = [WMICodeInline]
-
-    def wmi_codes_display(self, obj):
-        return ", ".join(obj.wmi_codes.values_list("code", flat=True))
-
-    wmi_codes_display.short_description = "WMI codes"
-
-
-@admin.register(EVModel)
-class EVModelAdmin(EntityModelAdmin):
-    fields = ("brand", "name")
-    list_display = ("name", "brand")
-
-
-admin.site.register(RFID, RFIDAdmin)
